# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Homium Compliance Portal API — templates, versions, saved sections, applications
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /applications/{applicationId}:
    get:
      operationId: getApplication
      tags: [intake]
      summary: Application with its intake state and pinned template
      parameters:
        - $ref: "#/components/parameters/ApplicationId"
      responses:
        "200":
          description: Application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
    patch:
      operationId: updateApplication
      tags: [intake]
      summary: Update portal-owned application fields
      description: >
        projectedClosingDate is a first-class portal-owned field (analyzer spec §8) — the
        hard-expiry clocks are judged against it. Null clears it.
      parameters:
        - $ref: "#/components/parameters/ApplicationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationUpdate"
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
...
  /applications/{applicationId}/blocks/{blockId}/variants:
    post:
      operationId: addVariant
      tags: [variants]
      summary: Add a variant to a set block on this application
      description: >
        Descriptor keys must match the template block's descriptorFields exactly;
        values non-empty; duplicate descriptor value-sets rejected (409). Label is
        built server-side; the id is minted once.
      parameters:
        - $ref: "#/components/parameters/ApplicationId"
        - $ref: "#/components/parameters/BlockId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [descriptor]
              properties:
                descriptor:
                  type: object
                  additionalProperties:
                    type: string
      responses:
        "201":
          description: The created variant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationVariant"
        "400":
          description: Block is not a set block, or descriptor keys/values invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
        "409":
          description: A variant with the same descriptor already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /applications/{applicationId}/blocks/{blockId}/variants/{variantId}:
    delete:
      operationId: deleteVariant
      tags: [variants]
      summary: Remove a variant (refused while uploads are tagged to it)
      parameters:
        - $ref: "#/components/parameters/ApplicationId"
        - $ref: "#/components/parameters/BlockId"
        - name: variantId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "409":
          description: Uploads still tagged to this variant — untag or delete them first
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
...
components:
  parameters:
...
    ApplicationId:
      name: applicationId
      in: path
      required: true
      schema:
        type: string
    BlockId:
      name: blockId
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    ApiMessage:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    ExpiryRule:
      type: object
      description: Null means no clock. staleness uses days; hard means valid through closing.
      required: [kind]
      properties:
        kind:
          type: string
          enum: [staleness, hard]
        days:
          type: integer
    Field:
      type: object
      required: [id, type, label]
      properties:
        id: { type: string }
        type:
          type: string
          enum: [text, number, date, select, yesno]
        label: { type: string }
        required: { type: boolean }
        options:
          type: array
          items: { type: string }
    Block:
      type: object
      description: kind=document uses document fields; kind=fields uses the fields array.
      required: [kind, id, name]
      properties:
        kind:
          type: string
          enum: [document, fields]
        id: { type: string }
        name: { type: string }
        docType:
          type: string
          description: >
            Optional analyzer taxonomy id (e.g. bank_statement). With it, classification is
            exact; without it, the analyzer falls back to name-similarity matching
            (analyzer spec §4). Never applicant-facing.
        formats:
          type: array
          items: { type: string }
        requirement:
          type: string
          enum: [required, required_alt, recommended, optional]
        criticality:
          type: string
          enum: [critical, standard, supporting]
        sourcing:
          type: string
          enum: [readily_available, constrained, scarce]
        multiPage: { type: boolean }
        expiry:
          oneOf:
            - $ref: "#/components/schemas/ExpiryRule"
            - type: "null"
        arity:
          type: string
          enum: [single, set]
          description: >
            "set" = the requirement is satisfied by N variants (bank accounts, persons, …),
            each with its own documents. Absent = "single" (today's behavior).
        variantConfig:
          $ref: "#/components/schemas/VariantConfig"
        analysisNote:
          type: string
          description: >
            Author-written expert guidance for the analyzer's satisfaction pass
            (e.g. "statements must be consecutive; balance carryover should match").
            Never applicant-facing.
        fields:
          type: array
          items:
            $ref: "#/components/schemas/Field"
    VariantConfig:
      type: object
      description: >
        Shape of a set block's variants. The template declares WHAT identifies a variant
        (descriptor field keys); the application supplies the actual variants per deal.
        Rules are data, never code (no conditional logic in v1).
      required: [variantNoun, descriptorFields, docsPerVariant]
      properties:
        variantNoun:
          type: string
          description: What one variant is called in the UI ("Bank account", "Person").
        descriptorFields:
          type: array
          description: Identity fields a variant must fill, as specific as possible.
          items:
            $ref: "#/components/schemas/DescriptorField"
        docsPerVariant:
          $ref: "#/components/schemas/DocsPerVariant"
    DescriptorField:
      type: object
      required: [key, label]
      properties:
        key: { type: string, description: "stable snake_case key (institution, account_last4, name, dob)" }
        label: { type: string, description: "UI label (\"Institution\", \"Account (last 4)\")" }
    DocsPerVariant:
      type: object
      required: [mode]
      properties:
        mode:
          type: string
          enum: [single, sequence]
          description: >
            single = exactly one document per variant (birth certificate);
            sequence = one or more documents (statements). One document never
            fails a sequence — counts are guidance, not gates.
        expectedCount:
          type: integer
          minimum: 1
          description: >
            Recommended number of documents per variant — guidance for review
            and the satisfaction pass, NEVER a hard limit or requirement.
            Document recency is governed by the block's existing expiry clock,
            not duplicated here.
        coverage:
          type: string
          enum: [consecutive_months]
          description: optional — the sequence must cover consecutive periods with no gaps.
    Subsection:
      type: object
      required: [id, name, blocks]
      properties:
        id: { type: string }
        name: { type: string }
        blocks:
          type: array
          items:
            $ref: "#/components/schemas/Block"
    Permission:
      type: object
      required: [role, view, upload]
      properties:
        role:
          type: string
          enum: [Applicant, Originator, Underwriter, Manager]
        view: { type: boolean }
        upload: { type: boolean }
    Section:
      type: object
      required: [id, name, owner, permissions, subsections]
      properties:
        id: { type: string }
        name: { type: string }
        owner:
          type: string
          enum: [Applicant, Originator, Escrow, Homium]
        permissions:
          type: array
          items:
            $ref: "#/components/schemas/Permission"
        subsections:
          type: array
          items:
            $ref: "#/components/schemas/Subsection"
    AlternativeGroup:
      type: object
      description: Satisfied when any one of [primary, ...satisfiedBy] is filed.
      required: [id, name, primary, satisfiedBy]
      properties:
        id: { type: string }
        name: { type: string }
        primary: { type: string }
        satisfiedBy:
          type: array
          items: { type: string }
    Template:
      type: object
      required: [template, version, status, program, alternatives, sections]
      properties:
        template: { type: string }
        version: { type: integer }
        status:
          type: string
          enum: [draft, active]
        program: { type: string }
        alternatives:
          type: array
          items:
            $ref: "#/components/schemas/AlternativeGroup"
        sections:
          type: array
          items:
            $ref: "#/components/schemas/Section"
...
    UploadedFile:
      type: object
      required: [filename, size, uploadedAt]
      properties:
        filename: { type: string }
        size: { type: integer }
        uploadedAt: { type: string }
        variantId:
          type: string
          description: >
            Set blocks only — which of the block's variants this file belongs to.
            Metadata only; object-store keys are unchanged.
        uploaderIp:
          type: string
          description: audit trail — client IP the upload arrived from
    ApplicationVariant:
      type: object
      description: >
        One real-world instance of a set block's requirement on THIS application
        ("Chase ····1234", "Jane Doe · 1990-04-24"). The template declares the
        descriptor shape; the application holds the actual variants. Id minted
        once, never re-minted — uploads and (later) approvals key on it.
      required: [id, descriptor, label, createdAt]
      properties:
        id: { type: string }
        descriptor:
          type: object
          description: descriptorField key -> value, keys exactly as the template block declares
          additionalProperties:
            type: string
        label: { type: string, description: "server-built display label (descriptor values joined)" }
        createdAt: { type: string }
...
    PageDecision:
      type: object
      description: A per-page pre-step decision inside the document approval flow.
      required: [page, decision]
      properties:
        page: { type: integer, description: "1-based packet page" }
        decision:
          type: string
          enum: [good, bad, flag_accepted]
          description: >
            good = page fine as-is; bad = page rejected; flag_accepted = page
            accepted despite flags/low scores — the flag stays on the record as
            a low-level note.
        note: { type: string }
    DocumentApproval:
      type: object
      description: >
        A per-DOCUMENT human approval (the filing decision) — the unit between
        page decisions and block verdicts. Tagged to a block and, for set
        blocks, a variant. approved/approved_incomplete materialize into the
        approved registry through the same seam as block accepts;
        approved_incomplete additionally requests a new version. Append-only on
        the application; a re-approval of the same pages supersedes via the
        registry, never deletes.
      required: [id, blockId, runId, pages, outcome, decidedBy, decidedAt]
      properties:
        id: { type: string }
        blockId: { type: string }
        variantId: { type: string, description: "set blocks — which variant this document files into" }
        runId: { type: string, description: "analyzer run whose grouping this approval is based on" }
        pages:
          type: array
          description: inclusive 1-based packet page range [first, last]
          items: { type: integer }
          minItems: 2
          maxItems: 2
        pageDecisions:
          type: array
          items:
            $ref: "#/components/schemas/PageDecision"
        outcome:
          type: string
          enum: [approved, approved_incomplete, rejected]
        decidedBy:
          type: string
          enum: [Originator, Underwriter, Manager]
        decidedAt: { type: string }
        approvedDocId: { type: string, description: "registry row created by materialization (approved outcomes only)" }
        pageRanges:
          type: array
          description: >
            For a document assembled from a human-accepted merge of
            NON-ADJACENT ranges: every inclusive [first,last] range that makes
            up the document, ascending, non-overlapping. `pages` remains the
            envelope [min,max] for old readers. Absent = single contiguous range.
          items:
            type: array
            items: { type: integer }
            minItems: 2
            maxItems: 2
...
    MergeResolution:
      type: object
      description: >
        The human decision on an analyzer merge recommendation between two run
        groups. Keyed on the application by "<runId>:p<f>-<l>|p<f>-<l>"
        (ranges sorted by first page). Latest decision wins — reversible.
      required: [runId, ranges, decision, decidedBy, decidedAt]
      properties:
        runId: { type: string }
        ranges:
          type: array
          description: the two inclusive page ranges the recommendation covers
          items:
            type: array
            items: { type: integer }
            minItems: 2
            maxItems: 2
          minItems: 2
          maxItems: 2
        decision:
          type: string
          enum: [merged, dismissed]
        decidedBy:
          type: string
          enum: [Originator, Underwriter, Manager]
        decidedAt: { type: string }
...
    ApplicationUpdate:
      type: object
      properties:
        projectedClosingDate:
          oneOf:
            - type: string
            - type: "null"
...
    Application:
      type: object
      required: [id, family, version, applicantName, createdAt, fieldValues, uploads, template]
      properties:
        id: { type: string }
        family: { type: string }
        version: { type: integer }
        applicantName: { type: string }
        createdAt: { type: string }
        fieldValues:
          type: object
          description: blockId -> field values map
          additionalProperties:
            type: object
            additionalProperties:
              type: string
        uploads:
          type: object
          description: blockId -> uploaded files
          additionalProperties:
            type: array
            items:
              $ref: "#/components/schemas/UploadedFile"
        projectedClosingDate: { type: string }
        verdicts:
          type: object
          description: blockId -> latest human verdict
          additionalProperties:
            $ref: "#/components/schemas/Verdict"
        packet:
          $ref: "#/components/schemas/PacketState"
        packetManifest:
          $ref: "#/components/schemas/PacketManifest"
        template:
          $ref: "#/components/schemas/Template"
        templateHistory:
          type: array
          description: Audit trail of template re-pins (who, when, vN→vN)
          items:
            $ref: "#/components/schemas/TemplateRepinEvent"
        materializationErrors:
          type: object
          description: >
            blockId -> last approval-materialization failure (loud, with reason).
            The verdict stands; the approved registry row is missing until a
            successful retry clears the entry. Portal-owned.
          additionalProperties:
            type: object
            required: [message, at]
            properties:
              message: { type: string }
              at: { type: string }
        documentApprovals:
          type: array
          description: >
            Per-document approvals from the filmstrip flow, append-only, newest
            first. The registry (approved-docs) is the materialized truth; this
            is the decision trail on the application. Portal-owned.
          items:
            $ref: "#/components/schemas/DocumentApproval"
        mergeResolutions:
          type: object
          description: >
            Merge-recommendation decisions, keyed "<runId>:p<f>-<l>|p<f>-<l>".
            A PENDING recommendation (no entry) gates approval of both groups.
            Portal-owned.
          additionalProperties:
            $ref: "#/components/schemas/MergeResolution"
        variants:
          type: object
          description: >
            blockId -> variants of that set block on this application. Intake-side
            data — variants and their uploads are NOT part of the application's
            satisfied requirements until a human accepts.
          additionalProperties:
            type: array
            items:
              $ref: "#/components/schemas/ApplicationVariant"
        manualPlacements:
          type: array
          description: Human filings of analyzer-unassigned page ranges (portal-owned)
          items:
            $ref: "#/components/schemas/ManualPlacement"
...
    TemplateRepinEvent:
      type: object
      required: [fromVersion, toVersion, decidedBy, decidedAt]
      properties:
        fromVersion: { type: integer }
        toVersion: { type: integer }
        decidedBy: { type: string }
        decidedAt: { type: string }
    Verdict:
      type: object
      description: >
        Human verdict on a document block. Portal-owned; the analyzer never writes these.
        documentDate/expiryDate are the confirmed dates the block's clocks run on.
      required: [verdict, datesEdited, decidedAt, decidedBy]
      properties:
        verdict:
          type: string
          enum: [accepted, new_version_requested]
        note: { type: string }
        documentDate: { type: string }
        expiryDate: { type: string }
        datesEdited: { type: boolean }
        decidedAt: { type: string }
        decidedBy:
          type: string
          enum: [Originator, Underwriter, Manager]
        runId: { type: string }
...
    ManualPlacement:
      type: object
      description: >
        A human decision about an unassigned page range: filed into a document block,
        or archived as not-relevant. Waits for the next analyzer run to confirm —
        manual placement always wins over analyzer suggestions.
      required: [pages, target, decidedBy, decidedAt]
      properties:
        pages:
          type: array
          description: inclusive 1-based page range [first, last]
          items:
            type: integer
          minItems: 2
          maxItems: 2
        target:
          type: string
          description: document blockId on the pinned template, or the literal "archive"
        note: { type: string }
        decidedBy:
          type: string
          enum: [Originator, Underwriter, Manager]
        decidedAt: { type: string }
        runId: { type: string }
...
    PacketMetadata:
      type: object
      description: >
        Metadata snapshot taken at pre-flight (analyzer spec §2.2). Display-only in v1 —
        metadata ANOMALY detection is analyzer-tier work, not pre-flight.
      properties:
        producer: { type: string }
        creator: { type: string }
        createdAt: { type: string }
        modifiedAt: { type: string }
    PacketThumbnail:
      type: object
      required: [page, reason]
      properties:
        page: { type: integer }
        reason:
          type: string
          description: Why pre-flight picked this page (blank, duplicate, lowest contrast, cleanest).
    PacketPreflight:
      type: object
      description: >
        Deterministic pre-flight report (analyzer spec §3) — no model calls, no image
        enhancement ("gate, don't retouch"). Checks actually run in v1: file validity,
        page count, metadata snapshot, per-page blank detection, per-page contrast,
        exact-duplicate pages, embedded-image DPI. Blur/skew scoring is analyzer-tier and
        intentionally NOT claimed here. Stored on the application as audit-trail material.
      required: [verdict, flags, estimateUsd, estimateMinutes, metadata, thumbnails]
      properties:
        verdict:
          type: string
          description: One plain-language line summarising the packet's pre-flight outcome.
        flags:
          type: array
          description: Plain-language red flags ("p.6 blank", "p.4 duplicate of p.3").
          items: { type: string }
        estimateUsd:
          type: number
          description: >
            FULL-pipeline estimate (parse + judge + deep scans) — informed consent before
            spend; the judge dominates at high page counts. Staff-facing only.
        estimateMinutes: { type: integer }
        metadata:
          $ref: "#/components/schemas/PacketMetadata"
        thumbnails:
          type: array
          description: 2 worst-scoring pages + 1 best, by deterministic per-page scores.
          items:
            $ref: "#/components/schemas/PacketThumbnail"
    PacketGateDecision:
      type: object
      required: [decision, decidedAt]
      properties:
        decision:
          type: string
          enum: [auto, confirmed, bypassed]
        decidedBy:
          type: string
          description: Absent when decision=auto; otherwise the signed-in staff profile.
        decidedAt: { type: string }
...
    PacketState:
      type: object
      description: >
        Portal-owned packet state machine — the staged C2 intake flow: preflight_running →
        gated → processing → report. Persisted server-side so the gate physically blocks;
        no client-side choreography can advance it. The spec §3 auto rule (<20 clean
        pages auto-proceed) is currently suspended: every packet gates so staff can
        pick the run's model plan before spend.
      required: [filename, sizeBytes, pages, sha256, uploadedAt, state]
      properties:
        filename: { type: string }
        sizeBytes: { type: integer }
        pages: { type: integer }
        sha256: { type: string }
        uploadedAt: { type: string }
        state:
          type: string
          enum: [preflight_running, gated, processing, report]
        preflight:
          $ref: "#/components/schemas/PacketPreflight"
        gate:
          $ref: "#/components/schemas/PacketGateDecision"
        lastRunError:
          type: string
          description: >
            Plain-language reason the last analyzer kick failed (packet reverted to gated).
            Cleared by the next successful upload, gate decision, or run ingest.
        uploaderIp:
          type: string
          description: audit trail — client IP the packet upload arrived from
        files:
          type: array
          description: >
            Multi-file provenance (assembled packets only): each source file's
            name and its [first,last] span in the packet's global page numbering.
            Absent on single-PDF uploads.
          items:
            type: object
            required: [filename, pages]
            properties:
              filename: { type: string }
              pages:
                type: array
                items: { type: integer }
                minItems: 2
                maxItems: 2
    PacketManifestFile:
      type: object
      required: [id, filename, sizeBytes, pages, flags, removed]
      properties:
        id: { type: string }
        filename: { type: string }
        sizeBytes: { type: integer }
        pages: { type: integer }
        flags:
          type: array
          description: deterministic per-file quality flags (same checks as pre-flight, no model calls)
          items: { type: string }
        removed: { type: boolean }
    PacketManifest:
      type: object
      description: >
        Multi-file intake staging (Phase 5 v1): the reviewable file list built before any
        analysis. Lives on the application until assemble replaces the packet; assembling
        marks it consumed. Removed files are simply skipped at assemble — no defer queue.
      required: [files, createdAt]
      properties:
        files:
          type: array
          items:
            $ref: "#/components/schemas/PacketManifestFile"
        createdAt: { type: string }
        assembledAt:
          type: string
          description: set when assemble consumed this manifest (staging bytes are gone)
...